Moving catalog_push logic to job
