+++ .ruby-gemset
@@ -1,0 +1,1 @@
+spree_retailops
