+++ Rakefile
@@ -1,11 +1,24 @@
 require 'bundler'
-Bundler::GemHelper.install_tasks
 
-require 'rspec/core/rake_task'
-require 'spree/testing_support/extension_rake'
+APP_RAKEFILE = File.expand_path("../spec/dummy/Rakefile", __FILE__)
+load 'rails/tasks/engine.rake'
 
 RSpec::Core::RakeTask.new
 
+Bundler::GemHelper.install_tasks
+
+Dir[File.join(File.dirname(__FILE__), 'tasks/**/*.rake')].each {|f| load f }
+
+require 'rspec/core'
+require 'rspec/core/rake_task'
+require 'spree/testing_support/extension_rake'
+
+desc "Run all specs in spec directory (excluding plugin specs)"
+RSpec::Core::RakeTask.new(:spec => 'app:db:test:prepare')
+
+task :default => :spec
+
+=begin
 task :default do
   if Dir["spec/dummy"].empty?
     Rake::Task[:test_app].invoke
@@ -19,3 +32,4 @@ task :test_app do
   ENV['LIB_NAME'] = 'spree_retailops'
   Rake::Task['extension:test_app'].invoke
 end
+=end
