+++ lib/spree_retailops.rb
@@ -2,3 +2,4 @@ require 'spree_core'
 require 'spree_retailops/engine'
 require 'spree/retailops/rop_stock_helper'
 require 'spree/retailops/rop_order_helper'
+require 'spree/api/retailops/catalog_helpers'
