+++ spree_retailops.gemspec
@@ -30,4 +30,7 @@ Gem::Specification.new do |s|
   s.add_development_dependency 'selenium-webdriver'
   s.add_development_dependency 'simplecov'
   s.add_development_dependency 'sqlite3'
+  s.add_development_dependency 'spree'
+  s.add_development_dependency 'activejob_backport'  # if RAILS_VERSION < "4.2"
+  s.add_development_dependency 'pry-byebug'
 end
